requirements.txt
# Apex AI Platform - Core Dependencies

# ============================================
# CORE PYTHON RUNTIME
# ============================================
python>=3.10

# ============================================
# WEB FRAMEWORK & API
# ============================================
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.10.6
python-multipart==0.0.18

# ============================================
# HTTP CLIENTS
# ============================================
requests==2.32.5
httpx==0.28.1

# ============================================
# DATA VALIDATION & PARSING
# ============================================
pyyaml==6.0.2
jsonschema==4.23.0
python-dotenv==1.0.1

# ============================================
# UI TESTING TOOLS
# ============================================
playwright==1.56.0
selenium==4.38.0

# ============================================
# API TESTING
# ============================================
# Already covered by requests/httpx

# ============================================
# TESTING FRAMEWORK
# ============================================
pytest==9.0.1
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# ============================================
# DATABASE & STORAGE
# ============================================
# SQLite is built-in with Python

# ============================================
# TEMPLATING & RENDERING
# ============================================
jinja2==3.1.5

# ============================================
# CLI & TERMINAL
# ============================================
rich==14.2.0
click==8.1.8

# ============================================
# UTILITIES
# ============================================
python-dateutil==2.9.0

# ============================================
# AI/LLM (Optional - Install in Phase 3)
# ============================================
# transformers>=4.30.0
# sentence-transformers>=2.2.0
# langchain>=0.1.0
# ollama>=0.1.0
# openai>=1.0.0
# anthropic>=0.20.0

# ============================================
# PERFORMANCE TESTING (Optional)
# ============================================
# locust>=2.20.0

# ============================================
# SECURITY TESTING (Optional)
# ============================================
# safety>=3.0.0
